reorganized user api files
